tests/wpt/web-platform-tests/preload/download-resources.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/preload/resources/preload_helper.js"></script>
<script>
    var t = async_test('Makes sure that preloaded resources are downloaded');
</script>
<link rel=preload href="resources/dummy.js" as=script>
<link rel=preload href="resources/dummy.css" as=style>
<link rel=preload href="resources/square.png" as=image>
<link rel=preload href="/media/CanvasTest.ttf" as=font crossorigin>
<link rel=preload href="/media/white.mp4" as=video>
<link rel=preload href="/media/sound_5.oga" as=audio>
<link rel=preload href="/media/foo.vtt" as=track>
<link rel=preload href="resources/dummy.xml?foo=bar" as=foobarxmlthing>
<link rel=preload href="resources/dummy.xml">
<body>
<script src="resources/dummy.js?pipe=trickle(d5)"></script>
<script>
    window.addEventListener("load", t.step_func(function() {
        verifyPreloadAndRTSupport()
        verifyNumberOfDownloads("resources/dummy.js", 1);
        verifyNumberOfDownloads("resources/dummy.css", 1);
        verifyNumberOfDownloads("/media/CanvasTest.ttf", 1);
        verifyNumberOfDownloads("/media/white.mp4", 1);
        verifyNumberOfDownloads("/media/sound_5.oga", 1);
        verifyNumberOfDownloads("/media/foo.vtt", 1);
        verifyNumberOfDownloads("resources/dummy.xml?foo=bar", 0);
        verifyNumberOfDownloads("resources/dummy.xml", 1);
        t.done();
    }));
</script>
</body>
